src/Web/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.2.1
cloud-init==25.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==3.1.1
Flask-Session==0.8.0
frozenlist==1.3.3
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
ms-identity-python @ git+https://github.com/azure-samples/ms-identity-python@834583646df315b79bc89f54808f4ae9c2a1ce59
msal==1.32.3
msgspec==0.19.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
openai==0.27.8
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dotenv==0.19.2
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
systemd-python==234
texttable==1.6.4
tqdm==4.65.0
Twisted==22.1.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.3
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
yarl==1.9.2
zipp==1.0.0
zope.interface==5.4.0
